refactor (WinningResultCalculator) : WinningResultCalculator를 당첨 집계 책임 클래스와 수익률 계산 책임 클래스로 분할
기존 WinningResultCalculator의 맴버변수가 각각 다른 메서드에서 활용되고 있었음, 단일 책임 불만족 냄새 해결
