+++ src/main/java/lotto/domain/WinningResultCalculator.java
@@ -1,47 +1,21 @@
 package lotto.domain;
 
+import javax.swing.plaf.PanelUI;
 import java.util.List;
 import java.util.Map;
 import java.util.TreeMap;
 
 public class WinningResultCalculator {
 
-    //TreeMap 으로 선언해야 합니다
-    Map<Rank, Integer> winningStatus = new TreeMap<>();
-    private int totalPrizeMoney = 0;
+    WinningStatusCalculator winningStatusCalculator = new WinningStatusCalculator();
+    ReturnRatioCalculator returnRatioCalculator = new ReturnRatioCalculator();
 
-    public double calculateReturnRatio(double expenses) {
-        calculateTotalPrizeMoney();
-        return (totalPrizeMoney / expenses) * 100;
+    public Map<Rank, Integer> generateWinningStatus(List<Lotto> lottos, WinningNumbers winningNumbers) {
+        return winningStatusCalculator.createWinningStatus(lottos, winningNumbers);
     }
 
-    public Map<Rank, Integer> createWinningStatus(List<Lotto> lottos, WinningNumbers winningNumbers) {
-
-        initWinningStaus();
-
-        for (Lotto lotto : lottos) {
-            Rank rank = Rank.calculateRank(lotto, winningNumbers);
-            winningStatus.put(rank, winningStatus.get(rank) + 1);
-        }
-
-        return winningStatus;
-    }
-
-    private void initWinningStaus(){
-        for (Rank rank : Rank.values()) {
-            winningStatus.put(rank, 0);
-        }
-    }
-
-    private int calculateTotalPrizeMoney() {
-
-        for (Rank rank : winningStatus.keySet()) {
-            for (int count = 0; count < winningStatus.get(rank); count++) {
-                totalPrizeMoney += Rank.getRankPrize(rank);
-            }
-        }
-
-        return totalPrizeMoney;
+    public double generateReturnRatio(double expenses, Map<Rank, Integer> winningStatus) {
+        return returnRatioCalculator.calculateReturnRatio(expenses, winningStatus);
     }
 
 }
